[sg noup] Link against shared libraries
Refreshing the preprocessed nng header is needed because we no longer
build against a predefined version.

Distutils makes it very hard to support customizable builds. Trying to
change the build system to support both, static and dynamic building
turned out to be pretty much impossible.

Therefore, all of this has works-for-me quality, can not be upstreamed.